Quickstart 4 - Add error handling for StopTheHost exception
The EF CLI throws a starts up the host to read database configuration and then throws a StopTheHost exception to stop the host from doing anything else. Check for that expected exception so that it is not logged as a fatal error.

See also https://github.com/dotnet/runtime/issues/60600.